Update network responses to support 64-bit Job IDs

Closes #28151

See merge request gitlab-org/gitlab-runner!3346

+++ common/network.go
@@ -154,7 +154,7 @@ type SessionInfo struct {
 type JobInfo struct {
 	Name        string `json:"name"`
 	Stage       string `json:"stage"`
-	ProjectID   int    `json:"project_id"`
+	ProjectID   int64  `json:"project_id"`
 	ProjectName string `json:"project_name"`
 }
 
@@ -332,7 +332,7 @@ type DependencyArtifactsFile struct {
 }
 
 type Dependency struct {
-	ID            int                     `json:"id"`
+	ID            int64                   `json:"id"`
 	Token         string                  `json:"token"`
 	Name          string                  `json:"name"`
 	ArtifactsFile DependencyArtifactsFile `json:"artifacts_file"`
@@ -346,7 +346,7 @@ type GitlabFeatures struct {
 }
 
 type JobResponse struct {
-	ID            int            `json:"id"`
+	ID            int64          `json:"id"`
 	Token         string         `json:"token"`
 	AllowGitFetch bool           `json:"allow_git_fetch"`
 	JobInfo       JobInfo        `json:"job_info"`
@@ -510,7 +510,7 @@ type JobTraceOutput struct {
 
 //nolint:lll
 type JobCredentials struct {
-	ID          int    `long:"id" env:"CI_JOB_ID" description:"The build ID to download and upload artifacts for"`
+	ID          int64  `long:"id" env:"CI_JOB_ID" description:"The build ID to download and upload artifacts for"`
 	Token       string `long:"token" env:"CI_JOB_TOKEN" required:"true" description:"Build token"`
 	URL         string `long:"url" env:"CI_SERVER_URL" required:"true" description:"GitLab CI URL"`
 	TLSCAFile   string `long:"tls-ca-file" env:"CI_SERVER_TLS_CA_FILE" description:"File containing the certificates to verify the peer when using HTTPS"`
@@ -539,7 +539,7 @@ func (j *JobCredentials) GetToken() string {
 }
 
 type UpdateJobInfo struct {
-	ID            int
+	ID            int64
 	State         JobState
 	FailureReason JobFailureReason
 	Output        JobTraceOutput
